Modified files: -------------- basic_toolkit/tests/eigenTest_1.sh   StRel20080125-patches only basic_toolkit/tests/eigenTest_2.sh basic_toolkit/tests/eigenTest_3.sh beamline/include/SeptumPropagators.h beamline/include/septum.h beamline/src/CF_rbend.cc beamline/src/CF_sbend.cc beamline/src/SeptumPropagators.cc beamline/src/TemplateInstantiations.cc beamline/src/bmlnElmnt.cc beamline/src/septum.cc beamline/src/thinMultipole.cc mxyzptlk/include/TJetEnvironment.h physics_toolkit/src/FPSolver.cc
Most changes are trivial, a few lines modified for compilation (on my
laptop) without the -fpermissive flag.  Exceptions are:

- septum.h, SeptumPropagators.h: incorporated Chong Shik Park's
class septum (from September, 2012).

- CF_sbend.cc, CF_rbend.cc: introduced one-line kludge that allows
the .peekAt function to be const while .OrbitLength remains non const.
The entire business of const or non-const for .OrbitLength is under
review.

- TemplateInstantiations.cc: instantiation of anonymous
boost::function1 with three arguments did not compile. I changed
"boost::function1" to "boost::function2" in those lines.

- SeptumPropagators.cc: incorporated Chong Shik Park's addition of
septum propagators, but did NOT change the behavior of thinSeptum
propagators.

Cherry-pick to branch StRel20080125-patches-cmake not (yet) done.

+++ physics_toolkit/src/FPSolver.cc
@@ -101,7 +101,7 @@ FPSolver::FPSolver( BmlPtr bml, int n )
 
   if( n <= 0 ||  (!bmLine_) ) {
     ostringstream uic;
-    uic << "Incorrect parameters: bml = " << (int) bml.get() 
+    uic << "Incorrect parameters: bml = " << (long) bml.get() 
         << " n = " << n;
     throw( GenericException( __FILE__, __LINE__, 
            "FPSolver::FPSolver( BmlPtr bml, int n )",
